+++ dune/gdt/assembler/local/codim1.hh
@@ -52,9 +52,9 @@ public:
     return {numTmpObjectsRequired_, localOperator_.numTmpObjectsRequired()};
   }
 
-  template <class TE, int TEd, int TEr, int TErC, class AE, int AEd, int AEr, int AErC, class TN, int TNd, int TNr,
-            int TNrC, class AN, int ANd, int ANr, int ANrC, class IntersectionType, class MEE, class MNN, class MEN,
-            class MNE, class R>
+  template <class TE, size_t TEd, size_t TEr, size_t TErC, class AE, size_t AEd, size_t AEr, size_t AErC, class TN,
+            size_t TNd, size_t TNr, size_t TNrC, class AN, size_t ANd, size_t ANr, size_t ANrC, class IntersectionType,
+            class MEE, class MNN, class MEN, class MNE, class R>
   void assembleLocal(const SpaceInterface<TE, TEd, TEr, TErC>& testSpaceEntity,
                      const SpaceInterface<AE, AEd, AEr, AErC>& ansatzSpaceEntity,
                      const SpaceInterface<TN, TNd, TNr, TNrC>& testSpaceNeighbor,
@@ -150,8 +150,8 @@ public:
     }
   } // void assembleLocal(...) const
 
-  template <class T, int Td, int Tr, int TrC, class A, int Ad, int Ar, int ArC, class IntersectionType, class M,
-            class R>
+  template <class T, size_t Td, size_t Tr, size_t TrC, class A, size_t Ad, size_t Ar, size_t ArC,
+            class IntersectionType, class M, class R>
   void assembleLocal(const SpaceInterface<T, Td, Tr, TrC>& testSpace, const SpaceInterface<A, Ad, Ar, ArC>& ansatzSpace,
                      const IntersectionType& intersection, Dune::Stuff::LA::MatrixInterface<M, R>& systemMatrix,
                      std::vector<std::vector<Dune::DynamicMatrix<R>>>& tmpLocalMatricesContainer,
@@ -204,8 +204,8 @@ public:
     return {numTmpObjectsRequired_, localOperator_.numTmpObjectsRequired()};
   }
 
-  template <class T, int Td, int Tr, int TrC, class A, int Ad, int Ar, int ArC, class IntersectionType, class M,
-            class R>
+  template <class T, size_t Td, size_t Tr, size_t TrC, class A, size_t Ad, size_t Ar, size_t ArC,
+            class IntersectionType, class M, class R>
   void assembleLocal(const SpaceInterface<T, Td, Tr, TrC>& testSpace, const SpaceInterface<A, Ad, Ar, ArC>& ansatzSpace,
                      const IntersectionType& intersection, Dune::Stuff::LA::MatrixInterface<M, R>& systemMatrix,
                      std::vector<std::vector<Dune::DynamicMatrix<R>>>& tmpLocalMatricesContainer,
@@ -284,7 +284,7 @@ public:
     return {numTmpObjectsRequired_, localFunctional_.numTmpObjectsRequired()};
   }
 
-  template <class T, int d, int r, int rC, class IntersectionType, class V, class R>
+  template <class T, size_t d, size_t r, size_t rC, class IntersectionType, class V, class R>
   void assembleLocal(const SpaceInterface<T, d, r, rC>& testSpace, const IntersectionType& intersection,
                      Dune::Stuff::LA::VectorInterface<V, R>& systemVector,
                      std::vector<std::vector<Dune::DynamicVector<R>>>& tmpLocalVectorsContainer,
