+++ dune/gdt/localevaluation/interface.hh
@@ -25,7 +25,7 @@ namespace LocalEvaluation {
  *  \tparam numArguments  The number of local bases.
  *  \note   All evaluations have to be copyable!
  */
-template <class Traits, int numArguments>
+template <class Traits, size_t numArguments>
 class Codim0Interface
 {
   static_assert(AlwaysFalse<Traits>::value, "There is no interface for this numArguments!");
@@ -43,7 +43,7 @@ public:
   typedef typename Traits::EntityType EntityType;
   typedef typename Traits::LocalfunctionTupleType LocalfunctionTupleType;
   typedef typename Traits::DomainFieldType DomainFieldType;
-  static const unsigned int dimDomain = Traits::dimDomain;
+  static const size_t dimDomain = Traits::dimDomain;
 
   LocalfunctionTupleType localFunctions(const EntityType& entity) const
   {
@@ -57,7 +57,7 @@ public:
    *  \tparam r   dimRange of the testBase
    *  \tparam rC  dimRangeRows of the testBase
    */
-  template <class R, int r, int rC>
+  template <class R, size_t r, size_t rC>
   size_t order(const LocalfunctionTupleType& localFunctions_in,
                const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, r, rC>& testBase) const
   {
@@ -72,7 +72,7 @@ public:
    *  \tparam rC  dimRangeRows of the testBase
    *  \attention ret is assumed to be zero!
    */
-  template <class R, int r, int rC>
+  template <class R, size_t r, size_t rC>
   void evaluate(const LocalfunctionTupleType& localFunctions_in,
                 const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, r, rC>& testBase,
                 const Dune::FieldVector<DomainFieldType, dimDomain>& localPoint, Dune::DynamicVector<R>& ret) const
@@ -93,7 +93,7 @@ public:
   typedef typename Traits::EntityType EntityType;
   typedef typename Traits::LocalfunctionTupleType LocalfunctionTupleType;
   typedef typename Traits::DomainFieldType DomainFieldType;
-  static const unsigned int dimDomain = Traits::dimDomain;
+  static const size_t dimDomain = Traits::dimDomain;
 
   LocalfunctionTupleType localFunctions(const EntityType& entity) const
   {
@@ -107,7 +107,7 @@ public:
    *  \tparam r{T,A}  dimRange of the {testBase,ansatzBase}
    *  \tparam rC{T,A} dimRangeRows of the {testBase,ansatzBase}
    */
-  template <class R, int rT, int rCT, int rA, int rCA>
+  template <class R, size_t rT, size_t rCT, size_t rA, size_t rCA>
   size_t
   order(const LocalfunctionTupleType& localFunctions,
         const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBase,
@@ -124,7 +124,7 @@ public:
    *  \tparam rC{L,T,A} dimRangeRows of the {localFunction,testBase,ansatzBase}
    *  \attention ret is assumed to be zero!
    */
-  template <class R, int rT, int rCT, int rA, int rCA>
+  template <class R, size_t rT, size_t rCT, size_t rA, size_t rCA>
   void evaluate(const LocalfunctionTupleType& localFunctions,
                 const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBase,
                 const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rA, rCA>& ansatzBase,
@@ -140,7 +140,7 @@ public:
  *  \tparam numArguments  The number of local bases.
  *  \note   All evaluations have to be copyable!
  */
-template <class Traits, int numArguments>
+template <class Traits, size_t numArguments>
 class Codim1Interface
 {
   static_assert(AlwaysFalse<Traits>::value, "There is no interface for this numArguments!");
@@ -158,7 +158,7 @@ public:
   typedef typename Traits::EntityType EntityType;
   typedef typename Traits::LocalfunctionTupleType LocalfunctionTupleType;
   typedef typename Traits::DomainFieldType DomainFieldType;
-  static const unsigned int dimDomain = Traits::dimDomain;
+  static const size_t dimDomain = Traits::dimDomain;
 
   LocalfunctionTupleType localFunctions(const EntityType& entity) const
   {
@@ -172,7 +172,7 @@ public:
    *  \tparam r   dimRange of the testBase
    *  \tparam rC  dimRangeRows of the testBase
    */
-  template <class R, int r, int rC>
+  template <class R, size_t r, size_t rC>
   size_t order(const LocalfunctionTupleType& localFunctions,
                const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, r, rC>& testBase) const
   {
@@ -188,7 +188,7 @@ public:
    *  \tparam rC                  dimRangeRows of the testBase
    *  \attention ret is assumed to be zero!
    */
-  template <class IntersectionType, class R, int r, int rC>
+  template <class IntersectionType, class R, size_t r, size_t rC>
   void evaluate(const LocalfunctionTupleType& localFunctions,
                 const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, r, rC>& testBase,
                 const IntersectionType& intersection,
@@ -210,7 +210,7 @@ public:
   typedef typename Traits::EntityType EntityType;
   typedef typename Traits::LocalfunctionTupleType LocalfunctionTupleType;
   typedef typename Traits::DomainFieldType DomainFieldType;
-  static const unsigned int dimDomain = Traits::dimDomain;
+  static const size_t dimDomain = Traits::dimDomain;
 
   LocalfunctionTupleType localFunctions(const EntityType& entity) const
   {
@@ -224,7 +224,7 @@ public:
    *  \tparam r{T,A}              dimRange of the {testBase,ansatzBase}
    *  \tparam rC{T,A}             dimRangeRows of the {testBase,ansatzBase}
    */
-  template <class R, int rT, int rCT, int rA, int rCA>
+  template <class R, size_t rT, size_t rCT, size_t rA, size_t rCA>
   size_t
   order(const LocalfunctionTupleType& localFunctions,
         const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBase,
@@ -242,7 +242,7 @@ public:
    *  \tparam rC{T,A}             dimRangeRows of the {testBase*,ansatzBase*}
    *  \attention ret is assumed to be zero!
    */
-  template <class IntersectionType, class R, int rT, int rCT, int rA, int rCA>
+  template <class IntersectionType, class R, size_t rT, size_t rCT, size_t rA, size_t rCA>
   void evaluate(const LocalfunctionTupleType& localFunctions,
                 const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBase,
                 const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rA, rCA>& ansatzBase,
@@ -265,7 +265,7 @@ public:
   typedef typename Traits::EntityType EntityType;
   typedef typename Traits::LocalfunctionTupleType LocalfunctionTupleType;
   typedef typename Traits::DomainFieldType DomainFieldType;
-  static const unsigned int dimDomain = Traits::dimDomain;
+  static const size_t dimDomain = Traits::dimDomain;
 
   LocalfunctionTupleType localFunctions(const EntityType& entity) const
   {
@@ -279,7 +279,7 @@ public:
    *  \tparam r{T,A}                dimRange of the {testBase*,ansatzBase*}
    *  \tparam rC{T,A}               dimRangeRows of the {testBase*,ansatzBase*}
    */
-  template <class R, int rT, int rCT, int rA, int rCA>
+  template <class R, size_t rT, size_t rCT, size_t rA, size_t rCA>
   size_t
   order(const LocalfunctionTupleType localFunctionsEntity, const LocalfunctionTupleType localFunctionsNeighbor,
         const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBaseEntity,
@@ -310,7 +310,7 @@ public:
    *  \tparam rC{T,A}               dimRangeRows of the {testBase*,ansatzBase*}
    *  \attention entityEntityRet, entityEntityRet, entityEntityRet and neighborEntityRet are assumed to be zero!
    */
-  template <class IntersectionType, class R, int rT, int rCT, int rA, int rCA>
+  template <class IntersectionType, class R, size_t rT, size_t rCT, size_t rA, size_t rCA>
   void evaluate(
       const LocalfunctionTupleType& localFunctionsEntity, const LocalfunctionTupleType& localFunctionsNeighbor,
       const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBaseEntity,
